requirements.txt
aiogram==3.4.1
supabase==2.3.4
requests==2.31.0
pillow==10.0.0
python-dotenv==1.0.0
tensorflow==2.12.0
beautifulsoup4==4.12.2
lxml==4.9.3
typing-extensions==4.7.1
h11==0.14.0
httpcore==0.17.3
urllib3==2.0.7
